filename parser moved to library
